Move dimension mapper and utilities to lib
