+++ packages/app/src/dimension-mapper/store.tsx
@@ -1,3 +1,4 @@
+import { type DimensionMapping, initDimMapping } from '@h5web/lib';
 import { type ArrayShape } from '@h5web/shared/hdf5-models';
 import {
   createContext,
@@ -9,8 +10,6 @@ import {
 import { createStore, type StoreApi, useStore } from 'zustand';
 
 import { areSameDims } from '../vis-packs/nexus/utils';
-import { type DimensionMapping } from './models';
-import { initDimMapping } from './utils';
 
 interface DimMappingState {
   dims: ArrayShape;
