+++ packages/app/src/vis-packs/core/complex/hooks.ts
@@ -1,11 +1,13 @@
-import { ComplexVisType } from '@h5web/lib';
+import {
+  ComplexVisType,
+  type DimensionMapping,
+  useSlicedDimsAndMapping,
+} from '@h5web/lib';
 import { type H5WebComplex } from '@h5web/shared/hdf5-models';
 import { type ComplexLineVisType } from '@h5web/shared/vis-models';
 import { type NdArray } from 'ndarray';
 import { useMemo } from 'react';
 
-import { type DimensionMapping } from '../../../dimension-mapper/models';
-import { useSlicedDimsAndMapping } from '../hooks';
 import { applyMapping, getBaseArray } from '../utils';
 import { getPhaseAmplitudeValues } from './utils';
