+++ packages/app/src/vis-packs/core/compound/CompoundVisContainer.tsx
@@ -1,3 +1,4 @@
+import { DimensionMapper, getSliceSelection } from '@h5web/lib';
 import {
   assertCompoundType,
   assertDataset,
@@ -5,14 +6,12 @@ import {
   assertPrintableCompoundType,
 } from '@h5web/shared/guards';
 
-import DimensionMapper from '../../../dimension-mapper/DimensionMapper';
 import { useValuesInCache } from '../../../dimension-mapper/hooks';
 import { useDimMappingState } from '../../../dimension-mapper/store';
 import visualizerStyles from '../../../visualizer/Visualizer.module.css';
 import { type VisContainerProps } from '../../models';
 import VisBoundary from '../../VisBoundary';
 import { useMatrixConfig } from '../matrix/config';
-import { getSliceSelection } from '../utils';
 import ValueFetcher from '../ValueFetcher';
 import MappedCompoundVis from './MappedCompoundVis';
