+++ packages/app/src/vis-packs/core/line/MappedLineVis.tsx
@@ -1,11 +1,14 @@
 import {
   type AxisParams,
+  type DimensionMapping,
+  getSliceSelection,
   type IgnoreValue,
   LineVis,
   useCombinedDomain,
   useDomain,
   useDomains,
   useSafeDomain,
+  useSlicedDimsAndMapping,
   useVisDomain,
 } from '@h5web/lib';
 import {
@@ -18,17 +21,15 @@ import {
 import { type AxisMapping } from '@h5web/shared/nexus-models';
 import { createPortal } from 'react-dom';
 
-import { type DimensionMapping } from '../../../dimension-mapper/models';
 import visualizerStyles from '../../../visualizer/Visualizer.module.css';
 import {
   useExportEntries,
   useMappedArray,
   useMappedArrays,
-  useSlicedDimsAndMapping,
   useToNumArray,
   useToNumArrays,
 } from '../hooks';
-import { DEFAULT_DOMAIN, formatNumLikeType, getSliceSelection } from '../utils';
+import { DEFAULT_DOMAIN, formatNumLikeType } from '../utils';
 import { type LineConfig } from './config';
 import LineToolbar from './LineToolbar';
 import { generateCsv } from './utils';
