+++ packages/app/src/vis-packs/core/matrix/MatrixVisContainer.tsx
@@ -1,16 +1,15 @@
+import { DimensionMapper, getSliceSelection } from '@h5web/lib';
 import {
   assertArrayShape,
   assertDataset,
   assertPrintableType,
 } from '@h5web/shared/guards';
 
-import DimensionMapper from '../../../dimension-mapper/DimensionMapper';
 import { useValuesInCache } from '../../../dimension-mapper/hooks';
 import { useDimMappingState } from '../../../dimension-mapper/store';
 import visualizerStyles from '../../../visualizer/Visualizer.module.css';
 import { type VisContainerProps } from '../../models';
 import VisBoundary from '../../VisBoundary';
-import { getSliceSelection } from '../utils';
 import ValueFetcher from '../ValueFetcher';
 import { useMatrixConfig } from './config';
 import MappedMatrixVis from './MappedMatrixVis';
