+++ packages/app/src/vis-packs/nexus/hooks.ts
@@ -1,3 +1,4 @@
+import { type DimensionMapping } from '@h5web/lib';
 import {
   type ComplexType,
   type GroupWithChildren,
@@ -5,7 +6,6 @@ import {
 } from '@h5web/shared/hdf5-models';
 
 import { useValuesInCache } from '../../dimension-mapper/hooks';
-import { type DimensionMapping } from '../../dimension-mapper/models';
 import { useDataContext } from '../../providers/DataProvider';
 import { type DatasetDef, type NxData } from './models';
 import {
